server/requirements.txt
fastapi
uvicorn
scipy
transformers
datasets
sentencepiece
huggingface_hub
accelerate
einops
hydra-core
hydra_colorlog
julius
num2words
numpy
tqdm
spacy
omegaconf
rich
soundfile
librosa
encodec
demucs
flashy
pesq
pystoi
torchdiffeq
torchmetrics
protobuf
